requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
backports.asyncio.runner==1.2.0
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.1
csvw==3.6.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
dacite==1.9.2
datasets==4.1.1
decorator==5.2.1
dill==0.4.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
einops==0.8.1
einx==0.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
exceptiongroup==1.3.0
fastapi==0.117.1
fastrtc==0.0.33
fastrtc-moonshine-onnx==20241016
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.9.23
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
google-crc32c==1.7.1
gradio==5.47.0
gradio_client==1.13.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
misaki==0.9.4
mistral_common==1.8.5
mlx==0.29.1
mlx-audio==0.2.5
mlx-lm==0.28.0
mlx-metal==0.29.1
mlx-vlm==0.3.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
murmurhash==1.0.13
networkx==3.4.2
nltk==3.9.1
num2words==0.5.14
numba==0.62.0
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.2
phonemizer-fork==3.3.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
sympy==1.14.0
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
pytesseract
pypdfium2
layoutparser[ocr]
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
torchvision
